Create SystemHandler classes which abstract out operations previously implemented in Application. Closes #44
