+++ src/main/java/pl/edu/agh/arbeit/tracker/Application.java
@@ -16,6 +16,8 @@ import com.sun.jna.ptr.IntByReference;
 import com.sun.jna.win32.StdCallLibrary;
 import pl.edu.agh.arbeit.tracker.events.ApplicationEvent;
 import pl.edu.agh.arbeit.tracker.events.EventType;
+import pl.edu.agh.arbeit.tracker.system.SystemHandler;
+import pl.edu.agh.arbeit.tracker.system.WindowsSystemHandler;
 import sun.reflect.generics.reflectiveObjects.NotImplementedException;
 
 /**
@@ -26,28 +28,22 @@ import sun.reflect.generics.reflectiveObjects.NotImplementedException;
 public class Application {
     private final String name;
     private final String programName;
+    private final SystemHandler handler;
 
     public Application(String name, String programName) {
         this.name = name;
         this.programName = programName;
+        this.handler = new WindowsSystemHandler();
+    }
+
+    public Application(String name, String programName, SystemHandler handler) {
+        this.name = name;
+        this.programName = programName;
+        this.handler = handler;
     }
 
     public boolean isRunning() {
-        try {
-            String line;
-            Process p = Runtime.getRuntime().exec(System.getenv("windir") + "\\system32\\" + "tasklist.exe");
-            BufferedReader input = new BufferedReader(new InputStreamReader(p.getInputStream()));
-            while ((line = input.readLine()) != null) {
-                if (line.contains(programName)) {
-                    return true;
-                }
-            }
-            input.close();
-            return false;
-        } catch (IOException e) {
-            e.printStackTrace();
-            return false;
-        }
+        return handler.getRunningApplications().contains(programName);
     }
     
     public ApplicationEvent getCurrentStateEvent(){
@@ -59,7 +55,7 @@ public class Application {
     }
 
     public boolean isActive() {
-        return programName.equals(getFocusedAppName());
+        return programName.equals(handler.getFocusedApplicationName());
     }
 
     public String getName() {
@@ -69,46 +65,7 @@ public class Application {
     public String getProgramName() {
         return programName;
     }
-
-
-    private interface Psapi extends StdCallLibrary {
-        Psapi INSTANCE = (Psapi) Native.loadLibrary("Psapi", Psapi.class);
-
-        WinDef.DWORD GetModuleBaseNameW(Pointer hProcess, Pointer hModule, byte[] lpBaseName, int nSize);
-    }
-
-    public static String getFocusedAppName() {
-        if (Platform.isWindows()) {
-            final int PROCESS_VM_READ = 0x0010;
-            final int PROCESS_QUERY_INFORMATION = 0x0400;
-            final User32 user32 = User32.INSTANCE;
-            final Kernel32 kernel32 = Kernel32.INSTANCE;
-            final Psapi psapi = Psapi.INSTANCE;
-            WinDef.HWND windowHandle = user32.GetForegroundWindow();
-            IntByReference pid = new IntByReference();
-            user32.GetWindowThreadProcessId(windowHandle, pid);
-            WinNT.HANDLE processHandle = kernel32.OpenProcess(PROCESS_VM_READ | PROCESS_QUERY_INFORMATION, true, pid.getValue());
-
-            byte[] filename = new byte[512];
-            Psapi.INSTANCE.GetModuleBaseNameW(processHandle.getPointer(), Pointer.NULL, filename, filename.length);
-            String focusedName = new String(filename);
-            focusedName = focusedName.replace("\0","");
-            if (focusedName.endsWith("wwahost.exe")) { // Metro App
-                // There is no stable API to get the current Metro app
-                // But you can guestimate the name form the current directory of the process
-                // To query this, see:
-                // http://stackoverflow.com/questions/16110936/read-other-process-current-directory-in-c-sharp
-                throw new NotImplementedException();
-            }
-            return focusedName;
-        } else {
-            //UNIX code
-            throw new NotImplementedException();
-        }
-
-    }
-
-    }
+}
