+++ src/main/java/pl/edu/agh/arbeit/tracker/system/SystemHandler.java
@@ -1,0 +1,9 @@
+package pl.edu.agh.arbeit.tracker.system;
+
+import java.util.Set;
+
+public interface SystemHandler {
+
+    Set<String> getRunningApplications();
+    String getFocusedApplicationName();
+}
